requirements/base.txt
# Base
pytz==2021.1
python-slugify==5.0.2
Pillow>=8.3.2
psycopg2==2.8.5 --no-binary psycopg2

# Django
django==3.2.6

# Django-crum
django-crum==0.7.9

# Environment
django-environ==0.7.0

# Django REST Framework
djangorestframework>=3.12.4
django-filter==21.1

# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==21.1.0

# Static files
whitenoise==5.3.0

# Celery
redis>=3.5.3
django-redis==5.0.0
celery==4.4.7
tornado>=5.0.0,<6.0.0
flower==0.9.5
